+++ src/SystemIntegration/TicketSystems/JiraTicketSystem.php
@@ -2,7 +2,7 @@
 
 namespace Becklyn\DeployMessageGenerator\SystemIntegration\TicketSystems;
 
-use Becklyn\DeployMessageGenerator\Config\DeployMessageGeneratorConfig;
+use Becklyn\DeployMessageGenerator\Config\DeployMessageGeneratorConfigurator;
 use Symfony\Component\Console\Style\SymfonyStyle;
 use Symfony\Component\HttpClient\HttpClient;
 use Symfony\Component\Notifier\Exception\TransportException;
@@ -15,15 +15,18 @@ use Symfony\Contracts\HttpClient\ResponseInterface;
 
 class JiraTicketSystem extends TicketSystem
 {
-    private String $deploymentFieldName;
-    private String $domain;
+    private string $domain;
     private string $jiraUser;
     private string $token;
+    private string $deploymentFieldKey;
 
 
+    /**
+     * @throws \Exception
+     */
     public function __construct (
         SymfonyStyle $io,
-        DeployMessageGeneratorConfig $config,
+        DeployMessageGeneratorConfigurator $config,
         string $deploymentFieldName,
         string $domain,
         string $jiraUser,
@@ -31,15 +34,11 @@ class JiraTicketSystem extends TicketSystem
     )
     {
         parent::__construct($io, $config);
-        $this->deploymentFieldName = $deploymentFieldName;
+
         $this->domain = $domain;
         $this->token = $token;
         $this->jiraUser = $jiraUser;
-
-        if (!$this->isDeploymentStatusFieldNameValid())
-        {
-            throw new \InvalidArgumentException("The field \"{$deploymentFieldName}\" does not exist for your jira installation.");
-        }
+        $this->deploymentFieldKey = $this->getDeploymentFieldKeyByFieldName($deploymentFieldName);
     }
 
 
@@ -81,9 +80,9 @@ class JiraTicketSystem extends TicketSystem
     {
         try
         {
-            $response = $this->sendRequest("GET", "https://{$this->domain}/rest/api/2/issue/{$id}?fields={$this->deploymentFieldName}");
+            $response = $this->sendRequest("GET", "https://{$this->domain}/rest/api/2/issue/{$id}?fields={$this->deploymentFieldKey}");
             $data = \json_decode($response->getContent(), true, 512, \JSON_THROW_ON_ERROR);
-            $field = $data["fields"][$this->deploymentFieldName];
+            $field = $data["fields"][$this->deploymentFieldKey];
             $status = "";
 
             if (!empty($field["value"]))
@@ -92,13 +91,15 @@ class JiraTicketSystem extends TicketSystem
             }
 
             return $status;
-        } catch (\Throwable $e) {
+        }
+        catch (\Throwable $e)
+        {
             throw new \Exception("Failed to make request to Jira", 1, $e);
         }
     }
 
 
-    protected function isDeploymentStatusFieldNameValid() : bool
+    private function getDeploymentFieldKeyByFieldName (string $deploymentFieldName) : string
     {
         try
         {
@@ -107,16 +108,18 @@ class JiraTicketSystem extends TicketSystem
 
             foreach ($fields as $field)
             {
-                if ($this->deploymentFieldName === ($field["name"] ?? null) || $this->deploymentFieldName === ($field["untranslatedName"] ?? null))
+                $name = $field["name"] ?? null;
+                $untranslatedName = $field["untranslatedName"] ?? null;
+
+                if ($deploymentFieldName === $name || $deploymentFieldName === $untranslatedName)
                 {
-                    $this->deploymentFieldName = $field["key"];
-                    return true;
+                    return $field["key"];
                 }
             }
 
-            return false;
+            throw new \InvalidArgumentException("The field \"{$deploymentFieldName}\" does not exist for your jira installation.");
         }
-        catch (\Throwable $e)
+        catch (\RuntimeException|RedirectionExceptionInterface|ClientExceptionInterface|TransportExceptionInterface|ServerExceptionInterface $e)
         {
             throw new \Exception("Failed to make request to Jira", 1, $e);
         }
@@ -128,7 +131,7 @@ class JiraTicketSystem extends TicketSystem
      * @throws ClientExceptionInterface
      * @throws TransportExceptionInterface
      * @throws ServerExceptionInterface
-     * @throws \Exception
+     * @throws \RuntimeException
      */
     protected function sendRequest (
         string $method,
@@ -153,7 +156,7 @@ class JiraTicketSystem extends TicketSystem
 
         if ("AUTHENTICATION_DENIED" === ($response->getHeaders(false)["X-Seraph-LoginReason"][0] ?? null))
         {
-            throw new \Exception("Cannot make API call to Jira. Captcha required");
+            throw new \RuntimeException("Cannot make API call to Jira. Captcha required");
         }
 
         return $response;
@@ -163,30 +166,24 @@ class JiraTicketSystem extends TicketSystem
     /**
      * @inheritDoc
      */
-    protected function setDeploymentStatus (string $id, ?string $deploymentStatus) : void
+    public function setDeploymentStatus (string $id, ?string $deploymentStatus) : void
     {
         try
         {
-            $status = null;
+            $status = !empty($deploymentStatus)
+                ? ["value" => $deploymentStatus]
+                : null;
 
-            if (!empty($deploymentStatus))
-            {
-                $status = [
-                    "value" => $deploymentStatus,
-                ];
-            }
-
-            $data = [
+            $response = $this->sendRequest("PUT", "https://{$this->domain}/rest/api/2/issue/{$id}", [
                 "json" => [
                     "fields" => [
-                        $this->deploymentFieldName => $status,
+                        $this->deploymentFieldKey => $status,
                     ],
                 ],
-            ];
+            ]);
 
-            $response = $this->sendRequest("PUT", "https://{$this->domain}/rest/api/2/issue/{$id}", $data);
-
-            if (204 !== $response->getStatusCode()) {
+            if (204 !== $response->getStatusCode())
+            {
                 throw new TransportException($response->getContent(false), $response, 1);
             }
         }
@@ -207,12 +204,80 @@ class JiraTicketSystem extends TicketSystem
 
 
     /**
+     * @throws DecodingExceptionInterface
+     * @throws ClientExceptionInterface
+     * @throws ServerExceptionInterface
+     * @throws RedirectionExceptionInterface
+     * @throws TransportExceptionInterface
+     * @throws \Exception
+     *
+     * @inheritDoc
+     */
+    public function generateDeployments (
+        array $context,
+        string $environment,
+        array $issueKeys,
+        array $urls,
+        string $commitRange
+    ) : JiraDeploymentResponse
+    {
+        if (0 === \count($issueKeys))
+        {
+            return JiraDeploymentResponse::fromErrors(["No Jira Issues have been provided."]);
+        }
+
+        if (0 === \count($urls))
+        {
+            return JiraDeploymentResponse::fromErrors(["No Deployment URL(s) have been provided."]);
+        }
+
+        $payload = [
+            "deployments" => $this->prepareDeploymentInformationPayload(
+                $environment,
+                $urls,
+                $issueKeys,
+                $commitRange
+            ),
+        ];
+
+        $jiraCloudId = $this->getJiraCloudId($context);
+        $jwtToken = $this->fetchJwtToken($context);
+
+        if (null === $jwtToken)
+        {
+            throw new \RuntimeException("Failed to connect to Atlassian API. Could not find existing 'JIRA_JWT_TOKEN' in environment, or generate a new JWT Token.");
+        }
+
+        $response = $this->sendDeploymentInformation($jiraCloudId, $payload, $jwtToken);
+
+        // If the previous request has failed due to an invalid/expired JWT Token, then generate a new Token and try again.
+        if (401 === $response->getStatusCode())
+        {
+            $response = $this->sendDeploymentInformation($jiraCloudId, $payload, $this->generateJiraJwt($context));
+        }
+
+        return JiraDeploymentResponse::fromResponse($response);
+    }
+
+
+    private function fetchJwtToken (array $context) : ?string
+    {
+        return $context["JIRA_JWT"] ?? $this->generateJiraJwt($context);
+    }
+
+
+    /**
      * Generate a JWT Token from the Jira API with Client Credentials.
      */
-    public function generateJiraJwt (array $context) : ?string
+    private function generateJiraJwt (array $context) : ?string
     {
-        $clientId = $context["JIRA_CLIENT_ID"];
-        $secret = $context["JIRA_CLIENT_SECRET"];
+        $clientId = $context["JIRA_CLIENT_ID"] ?? null;
+        $secret = $context["JIRA_CLIENT_SECRET"] ?? null;
+
+        if (null === $clientId || null === $secret)
+        {
+            throw new \RuntimeException("Could not generate JWT Token for Jira API without 'JIRA_CLIENT_ID' and 'JIRA_CLIENT_SECRET' present in the environment configuration.");
+        }
 
         $body = [
             "audience" => "api.atlassian.com",
@@ -225,7 +290,7 @@ class JiraTicketSystem extends TicketSystem
         {
             $tokenResponse = $this->sendRequest("POST", "https://api.atlassian.com/oauth/token", ["json" => $body]);
         }
-        catch (\Exception)
+        catch (\Exception $e)
         {
             return null;
         }
@@ -235,136 +300,116 @@ class JiraTicketSystem extends TicketSystem
 
 
     /**
-     * @throws DecodingExceptionInterface
-     * @throws ClientExceptionInterface
-     * @throws ServerExceptionInterface
-     * @throws RedirectionExceptionInterface
-     * @throws TransportExceptionInterface
-     * @throws \Exception
-     *
-     * @inheritDoc
+     * @param string[] $urls
+     * @param string[] $issueKeys
      */
-    public function generateDeployments (
-        array $context,
-        string $deploymentStatus,
+    private function prepareDeploymentInformationPayload (
+        string $environment,
+        array $urls,
         array $issueKeys,
-        string $url
-    ) : JiraDeploymentResponse
+        string $commitRange
+    ) : array
     {
-        $jiraDeploymentResponse = new JiraDeploymentResponse();
+        $pipelineIdPrefix = \uniqid("deployment-{$commitRange}-", true);
+        $environmentIdPrefix = \uniqid("environment-{$commitRange}-", true);
+        $deployments = [];
 
-        $pipelineId = \uniqid("deployment-", true);
-        $environmentId = \uniqid("environment-", true);
-
-        if (0 === \count($issueKeys))
+        foreach ($urls as $index => $url)
         {
-            $jiraDeploymentResponse->setErrors(["no Issue Keys where Provided."]);
-            return $jiraDeploymentResponse;
-        }
+            $hashedUrl = \sha1($url);
+            $pipelineId = "{$pipelineIdPrefix}-{$hashedUrl}";
+            $environmentId = "{$environmentIdPrefix}-{$hashedUrl}";
 
-        if (empty($context["JIRA_CLOUD_ID"]))
-        {
-            $response = $this
-                ->sendRequest("GET", "https://becklyn.atlassian.net/_edge/tenant_info")
-                ->toArray();
-
-            if (empty($response["cloudId"]))
-            {
-                throw new \RuntimeException("Could not get a Jira Cloud Id from 'https://becklyn.atlassian.net/_edge/tenant_info'.");
-            }
-
-            $cloudId = $response["cloudId"];
-        }
-        else
-        {
-            $cloudId = $context["JIRA_CLOUD_ID"];
-        }
-
-        $body = [
-            "deployments" => [[
-                "deploymentSequenceNumber" => 1,
-                "updateSequenceNumber" => 1,
+            $deployments[] = [
+                "deploymentSequenceNumber" => $index + 1,
+                "updateSequenceNumber" => $index + 1,
                 "issueKeys" => $issueKeys,
                 "displayName" => $pipelineId,
-                "description" => "Autogenerated Deployment with 'becklyn/deploy-message-generator'",
+                "description" => "Autogenerated Deployment with 'becklyn/deploy-message-generator' for Commit Range ‘{$commitRange}‘.",
                 "lastUpdated" => (new \DateTimeImmutable("now"))->format("Y-m-d\\TH:i:sP"),
                 "label" => $pipelineId,
                 "state" => "successful",
                 "url" => $url,
                 "pipeline" => [
                     "id" => $pipelineId,
-                    "displayName" => $deploymentStatus,
+                    "displayName" => $environment,
                     "url" => $url,
                 ],
                 "environment" => [
                     "id" => $environmentId,
-                    "displayName" => $deploymentStatus,
-                    "type" => $this->convertDeploymentStatus($deploymentStatus),
+                    "displayName" => $environment,
+                    "type" => $this->convertToJiraDeploymentEnvironment($environment),
                 ],
-            ]],
-        ];
-
-        if (!empty($context["JIRA_JWT"]))
-        {
-            $jwt = $context["JIRA_JWT"];
-
-            // try to send via env JWT
-            $response = $this->sendRequest(
-                "POST",
-                "https://api.atlassian.com/jira/deployments/0.1/cloud/{$cloudId}/bulk",
-                ["json" => $body],
-                "Bearer {$jwt}"
-            );
-
-            // generate a new JWT and try again
-            if (401 === $response->getStatusCode())
-            {
-                $jwt = $this->generateJiraJwt($context);
-
-                // abort if JWT generation fails
-                if (null === $jwt)
-                {
-                    throw new \Exception("Unable to generate a JWT.");
-                }
-
-                $response = $this->sendRequest(
-                    "POST",
-                    "https://api.atlassian.com/jira/deployments/0.1/cloud/{$cloudId}/bulk",
-                    ["json" => $body],
-                    "Bearer {$jwt}"
-                );
-            }
-
-            $jiraDeploymentResponse->setCode($response->getStatusCode());
-            $result = $response->toArray();
-
-            if (!empty($result["rejectedDeployments"]))
-            {
-                $errors = [];
-
-                foreach ($result["rejectedDeployments"] as $rejected)
-                {
-                    foreach ($rejected["errors"] as $error)
-                    {
-                        $errors[] = $error["message"];
-                    }
-                }
-
-                $jiraDeploymentResponse->setErrors($errors);
-            }
+            ];
         }
 
-        return $jiraDeploymentResponse;
+        return $deployments;
     }
 
 
-    private function convertDeploymentStatus (string $status) : string
+    /**
+     * @throws TransportExceptionInterface
+     * @throws ServerExceptionInterface
+     * @throws RedirectionExceptionInterface
+     * @throws ClientExceptionInterface
+     */
+    private function sendDeploymentInformation (
+        string $cloudId,
+        array $payload,
+        string $jwtToken
+    ) : ResponseInterface
     {
-        if ("Live" === $status)
+        return $this->sendRequest(
+            "POST",
+            "https://api.atlassian.com/jira/deployments/0.1/cloud/{$cloudId}/bulk",
+            [
+                "json" => $payload,
+            ],
+            "Bearer {$jwtToken}"
+        );
+    }
+
+
+    /**
+     * @throws RedirectionExceptionInterface
+     * @throws DecodingExceptionInterface
+     * @throws ClientExceptionInterface
+     * @throws TransportExceptionInterface
+     * @throws ServerExceptionInterface
+     */
+    private function getJiraCloudId (array $context) : string
+    {
+        $cloudId = $context["JIRA_CLOUD_ID"] ?? null;
+
+        if (null !== $cloudId)
         {
-            return "production";
+            return $cloudId;
         }
 
-        return $status;
+        $response = $this
+            ->sendRequest("GET", "https://{$this->domain}/_edge/tenant_info")
+            ->toArray();
+
+        $cloudId = $response["cloudId"] ?? null;
+
+        if (null === $cloudId)
+        {
+            throw new \RuntimeException("Could not get a Jira Cloud Id from 'https://{$this->domain}/_edge/tenant_info'.");
+        }
+
+        return $cloudId;
+    }
+
+
+    private function convertToJiraDeploymentEnvironment (string $environment) : string
+    {
+        switch ($environment)
+        {
+            case "Live":
+                return "production";
+
+            default:
+                return "staging";
+        }
     }
 }